+++ apps/service/src/app/(afterLogin)/home/page.tsx
@@ -1,6 +1,6 @@
 import { Suspense } from 'react';
 
-import UsageDashboard from '@service/components/UsageDashboard';
+import UsageDashboard from '@service/components/home/UsageDashboard';
 
 export default function Page() {
   return (
