+++ apps/service/src/components/common/NavBar.tsx
