+++ apps/service/src/components/common/ProgressBar.tsx
