+++ apps/service/src/components/home/CustomerItem.tsx
