+++ apps/service/src/components/home/ViewSegment.tsx
